Renaming class code files and format
Also insuring it compiles to all platforms properly
