+++ src/linkify/utils/class.js
@@ -1,4 +1,4 @@
-export default function inherits(parent, child, props={}) {
+export function inherits(parent, child, props={}) {
 	let extended = Object.create(parent.prototype);
 	for (var p in props) {
 		extended[p] = props[p];
